+++ sim/include/InfoSIRS.hh
@@ -34,9 +34,9 @@ namespace Models {
 
     void Init(po::variables_map& vm);
   
-    double getNodeEvents(eventList& events, State state,
+    unsigned int getNodeEvents(eventList& events, State state,
                          unsigned int nb) const;
-    double getEdgeEvents(eventList& events, State state,
+    unsigned int getEdgeEvents(eventList& events, State state,
                          unsigned int edge, State nbState, 
                          unsigned int nb) const;
 
@@ -66,14 +66,14 @@ namespace Models {
   
   private:
 
-    double gamma[2]; //!< Recovery rates.
-    double delta[2]; //!< Loss of immunity rates.
-    double beta[2][2]; //!< Infection rates.
-    double alpha; //!< Information transmission rate.
-    double lambda; //!< Rate of forgetting.
-    double omega; //!< Local infromation generation rate.
-    double nu; //!< Information generation rate over i-edges.
-    double sigma; //!< Ratio between informed and uninformed susceptibility.
+    unsigned int gamma[2]; //!< Recovery rates.
+    unsigned int delta[2]; //!< Loss of immunity rates.
+    unsigned int beta[2][2]; //!< Infection rates.
+    unsigned int alpha; //!< Information transmission rate.
+    unsigned int lambda; //!< Rate of forgetting.
+    unsigned int omega; //!< Local infromation generation rate.
+    unsigned int nu; //!< Information generation rate over i-edges.
+    double sigma; //!< Ratio between informed and uninformed susceptibility. 
   
   };
